requirements.txt
livekit-agents~=1.0
livekit-plugins-openai~=1.0
python-dotenv~=1.0
httpx~=0.27
